+++ webcomponents/charts/src/components/charts/deckdeckgo-bar-chart/deckdeckgo-bar-chart.tsx
@@ -6,9 +6,10 @@ import {BaseType, Selection} from 'd3-selection';
 import {scaleBand, scaleLinear} from 'd3-scale';
 import {max} from 'd3-array';
 import {Axis, axisBottom, axisLeft} from 'd3-axis';
+import {transition} from 'd3-transition';
 
 interface DeckdeckgoBarChartDataValue {
-  key: any;
+  key: string;
   value: number;
 }
 
@@ -26,62 +27,150 @@ export class DeckdeckgoBarChart implements DeckdeckgoChart {
 
   @Element() el: HTMLElement;
 
-  @Prop() width: number;
-  @Prop() height: number;
+  @Prop({mutable: true}) width: number;
+  @Prop({mutable: true}) height: number;
 
   @Prop() src: string;
   @Prop() separator: string = ';';
 
   @Prop() marginTop: number = 32;
-  @Prop() marginBottom: number = 32;
+  @Prop() marginBottom: number = 64;
   @Prop() marginLeft: number = 32;
   @Prop() marginRight: number = 32;
+
+  @Prop() animation: boolean = false;
+  @Prop() animationDuration: number = 1000;
+
+  private svg: Selection<BaseType, any, HTMLElement, any>;
+  private x0: any;
+  private x1: any;
+  private y: any;
+
+  private data: DeckdeckgoBarChartData[];
+
+  private barDataIndex: number = 0;
 
   async componentDidLoad() {
     await this.draw();
   }
 
-  @Watch('width')
-  @Watch('height')
   @Watch('src')
   async redraw() {
     await this.draw();
   }
 
   @Method()
-  draw(): Promise<void> {
+  draw(width?: number, height?: number): Promise<void> {
     return new Promise<void>(async (resolve) => {
+      if (width > 0) {
+        this.width = width;
+      }
+
+      if (height > 0) {
+        this.height = height;
+      }
+
       if (!this.width || !this.height || !this.src) {
         resolve();
         return;
       }
 
-      let svg: Selection<BaseType, any, HTMLElement, any> = DeckdeckgoChartUtils.initSvg(this.el, (this.width + this.marginLeft + this.marginRight), (this.height + this.marginTop + this.marginBottom));
-      svg = svg.append('g').attr('transform', 'translate(' + this.marginLeft + ',' + this.marginTop + ')');
+      this.barDataIndex = 0;
 
-      const data: DeckdeckgoBarChartData[] = await this.fetchData();
+      this.svg = DeckdeckgoChartUtils.initSvg(this.el, (this.width + this.marginLeft + this.marginRight), (this.height + this.marginTop + this.marginBottom));
+      this.svg = this.svg.append('g').attr('transform', 'translate(' + this.marginLeft + ',' + this.marginTop + ')');
 
-      const {x0, x1, y} = await this.initAxis(data);
+      this.data = await this.fetchData();
 
-      await this.drawAxis(svg, x0, y);
+      if (!this.data || this.data.length <= 0) {
+        resolve();
+        return;
+      }
 
-      await this.drawBars(svg, x0, x1, y, data);
+      await this.initAxis();
+
+      await this.drawAxis();
+
+      await this.drawBars(0, 0);
 
       resolve();
     });
   }
 
-  private drawAxis(svg: Selection<BaseType, any, HTMLElement, any>, x: any, y: any): Promise<void> {
-    return new Promise<void>((resolve) => {
-      const bottomAxis: Axis<any> = axisBottom(x);
-      const leftAxis: Axis<any> = axisLeft(y);
+  private async drawBars(index: number, animationDuration: number) {
+    if (this.animation) {
+      await this.drawAnimatedBars(index, animationDuration);
+    } else {
+      await this.drawInstantBars();
+    }
+  }
 
-      svg.append('g')
+  @Method()
+  async next() {
+    await this.prevNext(true);
+  }
+
+  @Method()
+  async prev() {
+    await this.prevNext(false);
+  }
+
+  private async prevNext(next: boolean) {
+    if (!this.animation) {
+      return;
+    }
+
+    if (!this.data || this.data.length <= 0) {
+      return;
+    }
+
+    if (next && this.barDataIndex + 1 < this.data.length) {
+      this.barDataIndex++;
+      await this.drawBars(this.barDataIndex, this.animationDuration);
+    } else if (!next && this.barDataIndex > 0) {
+      this.barDataIndex--;
+      await this.drawBars(this.barDataIndex, this.animationDuration);
+    }
+  }
+
+  @Method()
+  isBeginning(): Promise<boolean> {
+    return new Promise<boolean>((resolve) => {
+      if (!this.animation) {
+        resolve(true);
+        return;
+      }
+
+      resolve(this.barDataIndex === 0);
+    });
+  }
+
+  @Method()
+  isEnd(): Promise<boolean> {
+    return new Promise<boolean>((resolve) => {
+      if (!this.animation) {
+        resolve(true);
+        return;
+      }
+
+      resolve(this.barDataIndex === this.data.length - 1);
+    });
+  }
+
+  private drawAxis(): Promise<void> {
+    return new Promise<void>((resolve) => {
+      const bottomAxis: Axis<any> = axisBottom(this.x0);
+      const leftAxis: Axis<any> = axisLeft(this.y);
+
+      this.svg.append('g')
         .attr('class', 'axis axis-x')
         .attr('transform', 'translate(0,' + this.height + ')')
-        .call(bottomAxis);
+        .call(bottomAxis)
+        .selectAll('text')
+        .attr('transform', 'translate(-10,0)rotate(-45)')
+        .style('text-anchor', 'end');
 
-      svg.append('g')
+      this.svg.append('g')
         .attr('class', 'axis axis-y')
         .call(leftAxis);
 
@@ -89,41 +178,84 @@ export class DeckdeckgoBarChart implements DeckdeckgoChart {
     })
   }
 
-  private initAxis(data: DeckdeckgoBarChartData[]): Promise<any> {
-    return new Promise<any>((resolve) => {
-      let x0: any = scaleBand().rangeRound([0, this.width]).paddingInner(0.1);
-      ;
-      let x1: any = scaleBand().padding(0.05);
-      let y: any = scaleLinear().rangeRound([this.height, 0]);
+  private initAxis(): Promise<void> {
+    return new Promise<void>((resolve) => {
+      this.x0 = scaleBand().rangeRound([0, this.width]);
 
-      const categoriesNames = data.map((d) => {
-        return d.label;
-      });
-      const rateNames = data[0].values.map((d) => {
+      this.x1 = scaleBand().padding(0.05);
+      this.y = scaleLinear().rangeRound([this.height, 0]);
+
+      const xDomains = this.data[0].values.map((d) => {
         return d.key;
       });
 
-      x0.domain(categoriesNames);
-      x1.domain(rateNames).rangeRound([0, x0.bandwidth()]);
-      y.domain([0, max(data, (categorie) => {
-        return max(categorie.values, (d) => {
+      if (this.animation) {
+        this.initAnimatedAxisX(xDomains);
+      } else {
+        this.initInstantAxisX(xDomains);
+      }
+
+      this.y.domain([0, max(this.data, (category) => {
+        return max(category.values, (d) => {
           return d.value;
         });
       })]);
 
-      resolve({x0, x1, y});
+      resolve();
     });
   }
 
-  private drawBars(svg: Selection<BaseType, any, HTMLElement, any>, x0: any, x1: any, y: any, data: DeckdeckgoBarChartData[]): Promise<void> {
+  private initAnimatedAxisX(xDomains: string[]) {
+    this.x0.domain(xDomains).padding(0.05);
+  }
+
+  private initInstantAxisX(xDomains: string[]) {
+    const categoriesNames = this.data.map((d) => {
+      return d.label;
+    });
+
+    this.x0.paddingInner(0.1).domain(categoriesNames);
+    this.x1.domain(xDomains).rangeRound([0, this.x0.bandwidth()]);
+  }
+
+  private drawAnimatedBars(index: number, animationDuration: number): Promise<void> {
+    return new Promise<void>((resolve) => {
+      const t = transition();
+
+      const section: any = this.svg.selectAll('rect').data(this.data[index].values);
+
+      section
+        .enter()
+        .append('rect')
+        .merge(section)
+        .attr('style', (d) => {
+          return 'fill: var(--deckgo-chart-fill-color-' + d.key + '); fill-opacity: var(--deckgo-chart-fill-opacity-' + d.key + '); stroke: var(--deckgo-chart-stroke-' + d.key + '); stroke-width: var(--deckgo-chart-stroke-width-' + d.key + ')';
+        })
+        .transition(t).duration(animationDuration)
+        .attr('x', (d) => {
+          return this.x0(d.key);
+        })
+        .attr('y', (d) => {
+          return this.y(d.value);
+        })
+        .attr('width', this.x0.bandwidth())
+        .attr('height', (d) => {
+          return this.height - this.y(d.value);
+        });
+
+      resolve();
+    });
+  }
+
+  private drawInstantBars(): Promise<void> {
     return new Promise<void>((resolve) => {
 
-      svg.append('g')
+      this.svg.append('g')
         .selectAll('g')
-        .data(data)
+        .data(this.data)
         .enter().append('g')
         .attr('transform', (d) => {
-          return 'translate(' + x0(d.label) + ',0)';
+          return 'translate(' + this.x0(d.label) + ',0)';
         })
         .selectAll('rect')
         .data((d) => {
@@ -131,21 +263,21 @@ export class DeckdeckgoBarChart implements DeckdeckgoChart {
         })
         .enter().append('rect')
         .attr('x', (d) => {
-          return x1(d.key);
+          return this.x1(d.key);
         })
         .attr('y', (d) => {
-          return y(d.value);
+          return this.y(d.value);
         })
-        .attr('width', x1.bandwidth())
+        .attr('width', this.x1.bandwidth())
         .attr('height', (d) => {
-          return this.height - y(d.value);
+          return this.height - this.y(d.value);
         })
         .attr('style', (d) => {
           return 'fill: var(--deckgo-chart-fill-color-' + d.key + '); fill-opacity: var(--deckgo-chart-fill-opacity-' + d.key + '); stroke: var(--deckgo-chart-stroke-' + d.key + '); stroke-width: var(--deckgo-chart-stroke-width-' + d.key + ')';
         });
 
       resolve();
-    })
+    });
   }
 
   async fetchData(): Promise<DeckdeckgoBarChartData[]> {
@@ -181,9 +313,9 @@ export class DeckdeckgoBarChart implements DeckdeckgoChart {
             const tmp: number = parseInt(values[i]);
             if (!isNaN(tmp)) {
               dataValues.push({
-                key: 'bar' + i,
+                key: '' + i,
                 value: tmp
-              })
+              });
             }
           }